requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0
pyserial==3.5
python-engineio==4.8.0
python-socketio==5.10.0